First working LUA widget: Timer2
