+++ radio/src/lua/api_colorlcd.cpp
@@ -26,8 +26,8 @@
 
 #include "api_colorlcd.h"
 
-BitmapBuffer* luaLcdBuffer = nullptr;
-
+bool          luaLcdAllowed = false;
+BitmapBuffer* luaLcdBuffer  = nullptr;
 
 /*luadoc
 @function lcd.refresh()
@@ -252,17 +252,18 @@ static int luaLcdDrawTimer(lua_State *L)
   if (!luaLcdAllowed || !luaLcdBuffer)
     return 0;
 
-#if 0 // TODO
   int x = luaL_checkinteger(L, 1);
   int y = luaL_checkinteger(L, 2);
   int seconds = luaL_checkinteger(L, 3);
   unsigned int att = luaL_optunsigned(L, 4, 0);
 
-  if (att & SHADOWED) {
-    luaLcdBuffer->drawTimer(x+1, y+1, seconds, (att & 0xFFFF) | LEFT);
+  bool shadowed = att & SHADOWED;
+  att = (att & 0xFFFF) | COLOR(COLOR_VAL(att));
+
+  if (shadowed) {
+    drawTimer(luaLcdBuffer, x+1, y+1, seconds, (att & 0xFFFF) | LEFT);
   }
-  luaLcdBuffer->drawTimer(x, y, seconds, att | LEFT);
-#endif
+  drawTimer(luaLcdBuffer, x, y, seconds, att | LEFT);
 
   return 0;
 }
