+++ radio/src/lua/interface.cpp
@@ -29,6 +29,7 @@
 #include "sdcard.h"
 
 #if defined(LIBOPENUI)
+  #include "api_colorlcd.h"
   #include "libopenui.h"
 #else
   #include "libopenui/src/libopenui_file.h"
@@ -50,7 +51,6 @@ ScriptInputsOutputs scriptInputsOutputs[MAX_SCRIPTS];
 ScriptInternalData standaloneScript;
 uint16_t maxLuaInterval = 0;
 uint16_t maxLuaDuration = 0;
-bool luaLcdAllowed;
 uint8_t instructionsPercent = 0;
 char lua_warning_info[LUA_WARNING_INFO_LEN+1];
 struct our_longjmp * global_lj = 0;
